rename jdbc test class
