+++ catalogs/catalog-jdbc-doris/build.gradle.kts
@@ -69,6 +69,10 @@ tasks {
 
 tasks.test {
   val skipUTs = project.hasProperty("skipTests")
+  doFirst {
+    environment("GRAVITINO_CI_DORIS_DOCKER_IMAGE", "datastrato/gravitino-ci-doris:0.1.3")
+  }
+
   if (skipUTs) {
     // Only run integration tests
     include("**/integration/**")
@@ -80,10 +84,6 @@ tasks.test {
     exclude("**/integration/**")
   } else {
     dependsOn(tasks.jar)
-
-    doFirst {
-      environment("GRAVITINO_CI_DORIS_DOCKER_IMAGE", "datastrato/gravitino-ci-doris:0.1.3")
-    }
 
     val init = project.extra.get("initIntegrationTest") as (Test) -> Unit
     init(this)
