+++ catalogs/catalog-jdbc-mysql/src/test/java/com/datastrato/gravitino/catalog/mysql/operation/TestMysql.java
@@ -2,16 +2,14 @@
  * Copyright 2023 Datastrato Pvt Ltd.
  * This software is licensed under the Apache License version 2.
  */
-package com.datastrato.gravitino.catalog.mysql.integration.test;
+package com.datastrato.gravitino.catalog.mysql.operation;
 
+import com.datastrato.gravitino.catalog.jdbc.TestJdbc;
 import com.datastrato.gravitino.catalog.jdbc.config.JdbcConfig;
-import com.datastrato.gravitino.catalog.jdbc.integration.test.TestJdbcAbstractIT;
 import com.datastrato.gravitino.catalog.jdbc.utils.DataSourceUtils;
 import com.datastrato.gravitino.catalog.mysql.converter.MysqlColumnDefaultValueConverter;
 import com.datastrato.gravitino.catalog.mysql.converter.MysqlExceptionConverter;
 import com.datastrato.gravitino.catalog.mysql.converter.MysqlTypeConverter;
-import com.datastrato.gravitino.catalog.mysql.operation.MysqlDatabaseOperations;
-import com.datastrato.gravitino.catalog.mysql.operation.MysqlTableOperations;
 import com.datastrato.gravitino.integration.test.container.ContainerSuite;
 import com.datastrato.gravitino.integration.test.container.MySQLContainer;
 import com.datastrato.gravitino.integration.test.util.TestDatabaseName;
@@ -22,7 +20,7 @@ import java.util.Map;
 import javax.sql.DataSource;
 import org.junit.jupiter.api.BeforeAll;
 
-public class TestMysqlAbstractIT extends TestJdbcAbstractIT {
+public class TestMysql extends TestJdbc {
   public static final String defaultMysqlImageName = "mysql:8.0";
   protected static final ContainerSuite containerSuite = ContainerSuite.getInstance();
   protected static TestDatabaseName TEST_DB_NAME;
